:sparkles: New feature reporting page
Added a new reporting page to the Kesamokki application. This page includes
a chart that visualizes the number of users over time. The chart is created using Chart.js and is responsive to different screen sizes.
The reporting page is accessible via the top navigation bar and provides an overview of rental activity.
The charts are not working yet.
TODO: Fix the charts